basic implementation connection module
