--- brewblox_devcon_spark/connection.py
@@ -1,210 +1,0 @@
-"""
-Implements async serial connection.
-"""
-
-import asyncio
-from contextlib import suppress
-from subprocess import Popen
-from typing import Callable, Union
-
-from aiohttp import web
-from brewblox_service import brewblox_logger, features, repeater
-
-from brewblox_devcon_spark import (cbox_parser, connect_funcs, exceptions,
-                                   service_status, service_store)
-from brewblox_devcon_spark.models import (ControllerDescription,
-                                          DeviceDescription,
-                                          FirmwareDescription,
-                                          HandshakeMessage)
-
-MessageCallback_ = Callable[[str], None]
-
-LOGGER = brewblox_logger(__name__)
-
-BASE_RETRY_INTERVAL_S = 2
-MAX_RETRY_INTERVAL_S = 30
-CONNECT_RETRY_COUNT = 20
-
-WELCOME_PREFIX = 'BREWBLOX'
-UPDATER_PREFIX = 'FIRMWARE_UPDATER'
-CBOX_ERR_PREFIX = 'CBOXERROR'
-SETUP_MODE_PREFIX = 'SETUP_MODE'
-
-
-class SparkConnection(repeater.RepeaterFeature):
-
-    def __init__(self, app: web.Application):
-        super().__init__(app)
-
-        self._retry_count: int = 0
-        self._retry_interval: float = 0
-
-        self._proc: Popen = None
-        self._address: str = None
-        self._reader: asyncio.StreamReader = None
-        self._writer: asyncio.StreamWriter = None
-        self._parser: cbox_parser.ControlboxParser = None
-
-        self._data_callbacks = set()
-
-    def __str__(self):
-        return f'<{type(self).__name__} for {self._address}>'
-
-    @property
-    def connected(self) -> bool:
-        return bool(self._writer and not self._writer.is_closing())
-
-    @property
-    def data_callbacks(self) -> set[MessageCallback_]:
-        return self._data_callbacks
-
-    @property
-    def retry_interval(self) -> float:
-        if not self._retry_interval:
-            with service_store.fget(self.app).open() as config:
-                self._retry_interval = config.get('retry_interval', BASE_RETRY_INTERVAL_S)
-        return self._retry_interval
-
-    @retry_interval.setter
-    def retry_interval(self, value: float):
-        with service_store.fget(self.app).open() as config:
-            config['retry_interval'] = value
-        self._retry_interval = value
-
-    def reset_retry_interval(self):
-        self.retry_interval = BASE_RETRY_INTERVAL_S
-
-    def increase_retry_interval(self):
-        self.retry_interval = min(MAX_RETRY_INTERVAL_S, round(1.5 * self.retry_interval))
-
-    async def _on_event(self, msg: str):
-        if msg.startswith(WELCOME_PREFIX):
-            handshake = HandshakeMessage(*msg.split(','))
-            LOGGER.info(handshake)
-
-            desc = ControllerDescription(
-                system_version=handshake.system_version,
-                platform=handshake.platform,
-                reset_reason=handshake.reset_reason,
-                firmware=FirmwareDescription(
-                    firmware_version=handshake.firmware_version,
-                    proto_version=handshake.proto_version,
-                    firmware_date=handshake.firmware_date,
-                    proto_date=handshake.proto_date,
-                ),
-                device=DeviceDescription(
-                    device_id=handshake.device_id,
-                ),
-            )
-            service_status.set_acknowledged(self.app, desc)
-
-        elif msg.startswith(SETUP_MODE_PREFIX):
-            LOGGER.error('Controller entered listening mode. Exiting service now.')
-            raise web.GracefulExit()
-
-        else:
-            LOGGER.info(f'Spark event: `{msg}`')
-
-    async def _on_data(self, msg: str):
-        for cb in self._data_callbacks:
-            await cb(msg)
-
-    async def before_shutdown(self, app: web.Application):
-        await self.end()
-
-    async def run(self):
-        """Implements RepeaterFeature.run"""
-        try:
-            if self._retry_count >= CONNECT_RETRY_COUNT:
-                raise ConnectionAbortedError()
-            if self._retry_count == 1:
-                LOGGER.info('Retrying connection...')
-            if self._retry_count > 0:
-                await asyncio.sleep(self.retry_interval)
-
-            await service_status.wait_enabled(self.app)
-            result = await connect_funcs.connect(self.app)
-            self._proc = result.process
-            self._address = result.address
-            self._reader = result.reader
-            self._writer = result.writer
-            self._parser = cbox_parser.ControlboxParser()
-
-            service_status.set_connected(self.app, self._address)
-            self._retry_count = 0
-            self.reset_retry_interval()
-            LOGGER.info(f'{self} connected')
-
-            while self.connected:
-                # read() does not raise an exception when connection is closed
-                # connected status must be checked explicitly later
-                recv = await self._reader.read(100)
-
-                # read() returns empty if EOF received
-                if not recv:  # pragma: no cover
-                    raise ConnectionError('EOF received')
-
-                # Send to parser
-                self._parser.push(recv.decode())
-
-                # Drain parsed messages
-                for msg in self._parser.event_messages():
-                    await self._on_event(msg)
-                for msg in self._parser.data_messages():
-                    await self._on_data(msg)
-
-            raise ConnectionError('Connection closed')
-
-        except ConnectionAbortedError:
-            LOGGER.error('Connection aborted. Exiting now.')
-            self.increase_retry_interval()
-            raise web.GracefulExit()
-
-        except connect_funcs.DiscoveryAbortedError as ex:
-            LOGGER.error('Device discovery failed.')
-            if ex.reboot_required:
-                self._retry_count += 1
-            raise ex
-
-        except Exception:
-            self._retry_count += 1
-            raise
-
-        finally:
-            with suppress(Exception):
-                self._writer.close()
-                LOGGER.info(f'{self} closed stream writer')
-
-            with suppress(Exception):
-                self._proc.terminate()
-                LOGGER.info(f'{self} terminated subprocess')
-
-            service_status.set_disconnected(self.app)
-            self._proc = None
-            self._reader = None
-            self._writer = None
-            self._parser = None
-
-    async def write(self, msg: Union[str, bytes]):
-        if isinstance(msg, str):
-            msg = msg.encode()
-
-        if not self.connected:
-            raise exceptions.NotConnected(f'{self} not connected')
-
-        LOGGER.debug(f'{self} writing: {msg}')
-        self._writer.write(msg + b'\n')
-        await self._writer.drain()
-
-    async def start_reconnect(self):
-        # The run() function will handle cleanup, and then reconnect
-        if self.connected:
-            self._writer.close()
-
-
-def setup(app: web.Application):
-    features.add(app, SparkConnection(app))
-
-
-def fget(app: web.Application) -> SparkConnection:
-    return features.get(app, SparkConnection)
